renamed mask(), turned some more f64 into f32, version bump

+++ src/algebra/pos.rs
@@ -2,7 +2,7 @@ use std::ops::{Add, Mul};
 use crate::Signal;
 
 
-impl<N: Add<f64, Output = N> + Mul<f64, Output = N>> Signal<N> {
+impl<N: Add<f32, Output = N> + Mul<f32, Output = N>> Signal<N> {
     pub fn pos(self) -> Self {
         // shortcut to turn a -1;1 value into 0;1
         Signal {
